+++ NEWS.md
@@ -8,6 +8,10 @@
 
 3. New stage `sampling_poisson` to perform Poisson Disk Sampling
 
+### FIXES
+
+1. `sampling_*` stages respect the `filter` argument
+
 ### BREAKING CHANGES
 
 1. `classify_isolated_voxels()` renamed into `classify_with_ivf()` for consistency.